pyproject.toml
[project]
name = "castor-io"
version = "0.5.0"
description = "Minimalistic, embedded task manager."
readme = "README.md"
authors = [
    { name = "Alejandro Piad", email = "apiad@apiad.net" }
]
requires-python = ">=3.13"
dependencies = [
    "beaver-db>=0.16.8",
    "typer>=0.19.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["castor"]

[project.scripts]
castor = "castor.cli:app"
